+++ src/core/mod.rs
@@ -1,6 +1,8 @@
 #[cfg(feature = "modbus")]
 pub mod modbus;
 
+use core::{fmt, time::Duration, result::Result};
+
 /// (Thermodynamic) Temperature.
 #[derive(Clone, Copy, Debug, PartialEq, PartialOrd)]
 #[repr(transparent)]
@@ -16,15 +18,9 @@ impl Temperature {
     }
 }
 
-impl From<f64> for Temperature {
-    fn from(from: f64) -> Self {
-        Temperature(from)
-    }
-}
-
-impl From<Temperature> for f64 {
-    fn from(from: Temperature) -> Self {
-        from.0
+impl fmt::Display for Temperature {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{} °C", self.to_degree_celsius())
     }
 }
 
@@ -63,6 +59,12 @@ impl VolumetricWaterContent {
     }
 }
 
+impl fmt::Display for VolumetricWaterContent {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{} %", self.to_percent())
+    }
+}
+
 /// Relative permittivity or dielectric constant (DK).
 #[derive(Clone, Copy, Debug, PartialEq, PartialOrd)]
 #[repr(transparent)]
@@ -88,6 +90,59 @@ impl RelativePermittivity {
     pub fn is_valid(self) -> bool {
         self >= Self::min()
     }
+}
+
+impl fmt::Display for RelativePermittivity {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{} %", self.to_ratio())
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, PartialOrd, Ord)]
+pub struct RawCounts(u16);
+
+impl From<u16> for RawCounts {
+    fn from(from: u16) -> Self {
+        RawCounts(from)
+    }
+}
+
+impl From<RawCounts> for u16 {
+    fn from(from: RawCounts) -> Self {
+        from.0
+    }
+}
+
+impl fmt::Display for RawCounts {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Blocking interface that exposes the generic capabilities of the
+/// TRUEBNER SMT100 Soil Moisture Sensor.
+pub trait Capabilities {
+    type ReadError;
+
+    /// Measure the current temperature in the range from -40°C to +80°C
+    /// (analog version from -40°C to +60°C).
+    fn read_temperature(&self, timeout: Option<Duration>) -> Result<Temperature, Self::ReadError>;
+
+    /// Measure the current water content of the medium (soil) around the sensor
+    /// in the range from 0% to 60% (up to 100% with limited accuracy).
+    fn read_water_content(
+        &self,
+        timeout: Option<Duration>,
+    ) -> Result<VolumetricWaterContent, Self::ReadError>;
+
+    /// Measure the current (relative) permittivity of the medium around the sensor.
+    fn read_permittivity(
+        &self,
+        timeout: Option<Duration>,
+    ) -> Result<RelativePermittivity, Self::ReadError>;
+
+    /// Retrieve the current raw and uncalibrated signal of the sensor.
+    fn read_raw_counts(&self, timeout: Option<Duration>) -> Result<RawCounts, Self::ReadError>;
 }
 
 #[cfg(test)]