+++ src/core/modbus/mod.rs
@@ -3,38 +3,54 @@ use super::*;
 #[cfg(feature = "rtu")]
 pub mod rtu;
 
-use byteorder::{ByteOrder, BigEndian};
-use core::result::Result;
+use core::{fmt, convert::TryInto};
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum DecodeError {
+    InsufficientInput,
     InvalidInput,
     InvalidData,
 }
 
+impl fmt::Display for DecodeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        use DecodeError::*;
+        match self {
+            InsufficientInput => write!(f, "Insufficient input"),
+            InvalidInput => write!(f, "Invalid input"),
+            InvalidData => write!(f, "Invalid data"),
+        }
+
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for DecodeError {}
+
 pub type DecodeResult<T> = Result<T, DecodeError>;
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct TemperatureRaw(pub u16);
-
-impl From<TemperatureRaw> for Temperature {
-    fn from(from: TemperatureRaw) -> Self {
-        let degree_celsius = f64::from(i32::from(from.0) - 10000i32) / 100f64;
-        Self::from_degree_celsius(degree_celsius)
+fn decode_be_u16_from_bytes(input: &[u8]) -> DecodeResult<(u16, &[u8])> {
+    if input.len() < std::mem::size_of::<u16>() {
+        return Err(DecodeError::InsufficientInput);
+    }
+    let (head, rest) = input.split_at(std::mem::size_of::<u16>());
+    if let Ok(bytes) = head.try_into() {
+        Ok((u16::from_be_bytes(bytes), rest))
+    } else {
+        Err(DecodeError::InvalidInput)
     }
 }
 
 pub const TEMPERATURE_REG_START: u16 = 0x0000;
 pub const TEMPERATURE_REG_COUNT: u16 = 0x0001;
 
-const TEMPERATURE_BYTES_LEN: usize = TEMPERATURE_REG_COUNT as usize * 2;
+pub fn decode_temperature_from_u16(input: u16) -> DecodeResult<Temperature> {
+    let degree_celsius = f64::from(i32::from(input) - 10000i32) / 100f64;
+    Ok(Temperature::from_degree_celsius(degree_celsius))
+}
 
-pub fn decode_temperature_bytes(bytes: &[u8]) -> DecodeResult<(Temperature, &[u8])> {
-    if bytes.len() < TEMPERATURE_BYTES_LEN {
-        return Err(DecodeError::InvalidInput);
-    }
-    let raw = BigEndian::read_u16(bytes);
-    Ok((TemperatureRaw(raw).into(), &bytes[TEMPERATURE_BYTES_LEN..]))
+pub fn decode_temperature_from_bytes(input: &[u8]) -> DecodeResult<(Temperature, &[u8])> {
+    decode_be_u16_from_bytes(input).and_then(|(val, rest)| Ok((decode_temperature_from_u16(val)?, rest)))
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -50,18 +66,18 @@ impl From<VolumetricWaterContentRaw> for VolumetricWaterContent {
 pub const WATER_CONTENT_REG_START: u16 = 0x0001;
 pub const WATER_CONTENT_REG_COUNT: u16 = 0x0001;
 
-const WATER_CONTENT_BYTES_LEN: usize = WATER_CONTENT_REG_COUNT as usize * 2;
+pub fn decode_water_content_from_u16(input: u16) -> DecodeResult<VolumetricWaterContent> {
+    let percent = f64::from(input) / 100f64;
+    let res = VolumetricWaterContent::from_percent(percent);
+    if res.is_valid() {
+        Ok(res)
+    } else {
+        Err(DecodeError::InvalidData)
+    }
+}
 
-pub fn decode_water_content_bytes(bytes: &[u8]) -> DecodeResult<(VolumetricWaterContent, &[u8])> {
-    if bytes.len() < WATER_CONTENT_BYTES_LEN {
-        return Err(DecodeError::InvalidInput);
-    }
-    let raw = BigEndian::read_u16(bytes);
-    let val: VolumetricWaterContent = VolumetricWaterContentRaw(raw).into();
-    if !val.is_valid() {
-        return Err(DecodeError::InvalidData);
-    }
-    Ok((val, &bytes[WATER_CONTENT_BYTES_LEN..]))
+pub fn decode_water_content_from_bytes(input: &[u8]) -> DecodeResult<(VolumetricWaterContent, &[u8])> {
+    decode_be_u16_from_bytes(input).and_then(|(val, rest)| Ok((decode_water_content_from_u16(val)?, rest)))
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -77,31 +93,31 @@ impl From<RelativePermittivityRaw> for RelativePermittivity {
 pub const PERMITTIVITY_REG_START: u16 = 0x0002;
 pub const PERMITTIVITY_REG_COUNT: u16 = 0x0001;
 
-const PERMITTIVITY_BYTES_LEN: usize = PERMITTIVITY_REG_COUNT as usize * 2;
+pub fn decode_permittivity_from_u16(input: u16) -> DecodeResult<RelativePermittivity> {
+    let ratio = f64::from(input) / 100f64;
+    let res = RelativePermittivity::from_ratio(ratio);
+    if res.is_valid() {
+        Ok(res)
+    } else {
+        Err(DecodeError::InvalidData)
+    }
+}
 
-pub fn decode_permittivity_bytes(bytes: &[u8]) -> DecodeResult<(RelativePermittivity, &[u8])> {
-    if bytes.len() < PERMITTIVITY_BYTES_LEN {
-        return Err(DecodeError::InvalidInput);
-    }
-    let raw = BigEndian::read_u16(bytes);
-    let val: RelativePermittivity = RelativePermittivityRaw(raw).into();
-    if !val.is_valid() {
-        return Err(DecodeError::InvalidData);
-    }
-    Ok((val, &bytes[PERMITTIVITY_BYTES_LEN..]))
+pub fn decode_permittivity_from_bytes(input: &[u8]) -> DecodeResult<(RelativePermittivity, &[u8])> {
+    decode_be_u16_from_bytes(input).and_then(|(val, rest)| Ok((decode_permittivity_from_u16(val)?, rest)))
 }
 
 pub const RAW_COUNTS_REG_START: u16 = 0x0003;
 pub const RAW_COUNTS_REG_COUNT: u16 = 0x0001;
 
-const RAW_COUNTS_BYTES_LEN: usize = RAW_COUNTS_REG_COUNT as usize * 2;
+#[inline]
+pub fn decode_raw_counts_from_u16(input: u16) -> DecodeResult<RawCounts> {
+    Ok(input.into())
+}
 
-pub fn decode_raw_counts(bytes: &[u8]) -> DecodeResult<(u16, &[u8])> {
-    if bytes.len() < RAW_COUNTS_BYTES_LEN {
-        return Err(DecodeError::InvalidInput);
-    }
-    let val = BigEndian::read_u16(bytes);
-    Ok((val, &bytes[RAW_COUNTS_BYTES_LEN..]))
+#[inline]
+pub fn decode_raw_counts_from_bytes(input: &[u8]) -> DecodeResult<(RawCounts, &[u8])> {
+    decode_be_u16_from_bytes(input).and_then(|(val, rest)| Ok((decode_raw_counts_from_u16(val)?, rest)))
 }
 
 pub const BROADCAST_SLAVE_ADDR: u8 = 0xFD;
@@ -115,23 +131,23 @@ mod tests {
     fn decode_temperature() {
         assert_eq!(
             Temperature::from_degree_celsius(-40.0),
-            decode_temperature_bytes(&[0x17, 0x70]).unwrap().0
+            decode_temperature_from_bytes(&[0x17, 0x70]).unwrap().0
         );
         assert_eq!(
             Temperature::from_degree_celsius(0.0),
-            decode_temperature_bytes(&[0x27, 0x10]).unwrap().0
+            decode_temperature_from_bytes(&[0x27, 0x10]).unwrap().0
         );
         assert_eq!(
             Temperature::from_degree_celsius(27.97),
-            decode_temperature_bytes(&[0x31, 0xFD]).unwrap().0
+            decode_temperature_from_bytes(&[0x31, 0xFD]).unwrap().0
         );
         assert_eq!(
             Temperature::from_degree_celsius(60.0),
-            decode_temperature_bytes(&[0x3E, 0x80]).unwrap().0
+            decode_temperature_from_bytes(&[0x3E, 0x80]).unwrap().0
         );
         assert_eq!(
             Temperature::from_degree_celsius(80.0),
-            decode_temperature_bytes(&[0x46, 0x50]).unwrap().0
+            decode_temperature_from_bytes(&[0x46, 0x50]).unwrap().0
         );
     }
 
@@ -140,19 +156,19 @@ mod tests {
         // Valid range
         assert_eq!(
             VolumetricWaterContent::from_percent(0.0),
-            decode_water_content_bytes(&[0x00, 0x00]).unwrap().0
+            decode_water_content_from_bytes(&[0x00, 0x00]).unwrap().0
         );
         assert_eq!(
             VolumetricWaterContent::from_percent(34.4),
-            decode_water_content_bytes(&[0x0D, 0x70]).unwrap().0
+            decode_water_content_from_bytes(&[0x0D, 0x70]).unwrap().0
         );
         assert_eq!(
             VolumetricWaterContent::from_percent(100.0),
-            decode_water_content_bytes(&[0x27, 0x10]).unwrap().0
+            decode_water_content_from_bytes(&[0x27, 0x10]).unwrap().0
         );
         // Invalid range
-        assert!(decode_water_content_bytes(&[0x27, 0x11]).is_err());
-        assert!(decode_water_content_bytes(&[0xFF, 0xFF]).is_err());
+        assert!(decode_water_content_from_bytes(&[0x27, 0x11]).is_err());
+        assert!(decode_water_content_from_bytes(&[0xFF, 0xFF]).is_err());
     }
 
     #[test]
@@ -160,14 +176,14 @@ mod tests {
         // Valid range
         assert_eq!(
             RelativePermittivity::from_ratio(1.0),
-            decode_permittivity_bytes(&[0x00, 0x64]).unwrap().0
+            decode_permittivity_from_bytes(&[0x00, 0x64]).unwrap().0
         );
         assert_eq!(
             RelativePermittivity::from_ratio(15.2),
-            decode_permittivity_bytes(&[0x05, 0xF0]).unwrap().0
+            decode_permittivity_from_bytes(&[0x05, 0xF0]).unwrap().0
         );
         // Invalid range
-        assert!(decode_permittivity_bytes(&[0x00, 0x00]).is_err());
-        assert!(decode_permittivity_bytes(&[0x00, 0x63]).is_err());
+        assert!(decode_permittivity_from_bytes(&[0x00, 0x00]).is_err());
+        assert!(decode_permittivity_from_bytes(&[0x00, 0x63]).is_err());
     }
 }