+++ CHANGELOG.md
@@ -8,13 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added *no_std* features *modbus*, *rtu*, and *modbus-rtu*
+- Added `[no_std]` core module with optional features `modbus` and `rtu`
+- Added various `[no_std]` low-level *Modbus* decoding functions
+- Added a `[no_std]` blocking `Capabilities` trait
 
 ### Changed
 
-- Renamed feature *modbus-rtu* as *tokio-modbus-rtu*
+- Renamed feature `modbus-rtu` as `tokio-modbus-rtu`
+- Renamed feature `mock` as `tokio-mock`
+- Read timeout on the non-blocking `Capabilities` trait has become optional
 
 ### Removed
+
+- Removed dependency on the `byteorder` crate
+- Removed all *Newtypes* for `...Raw` measurements
 
 ## [0.2.1] - 2019-05-21